requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
pycodestyle==2.10.0
pymongo==4.3.3
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
yarl==1.8.2
